make sizer adjustment functions public
